# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API du module de billeterie automatique",
        "version": "1.0.0"
    },
    "paths": {
        "/api/all_clients": {
            "get": {
                "tags": [
                    "The_client"
                ],
                "summary": "Afficher la liste des client",
                "description": "Renvoyer la liste de tous les clients de marchandises",
                "operationId": "ac04554a38a71a01f618a87ca698d061",
                "responses": {
                    "200": {
                        "description": "Les clients de colis ont été trouvé et leurs informations sont retournées",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun client de marchandise dans la base de données ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client_show/{id}": {
            "get": {
                "tags": [
                    "The_client"
                ],
                "summary": "Afficher les infos d'un client",
                "description": "Renvoyer les infos d'un client",
                "operationId": "65931d721ab2d0e4b505259c6170bfe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant d'un client à voir ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " trouvé  !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun  ne correspond à l'identifiant saisi",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client_store": {
            "post": {
                "tags": [
                    "The_client"
                ],
                "summary": "Enregistrer un nouveau client",
                "operationId": "6556db910ef366718b28ca90fff60453",
                "parameters": [
                    {
                        "name": "nom_prenom",
                        "in": "query",
                        "description": "Nom et prénom du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telephone",
                        "in": "query",
                        "description": "Téléphone du client.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email du client.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " enregistre avec succes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Echec d'enregistrement de l'.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "402": {
                        "description": "Un client avec ces informations existe déja",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/client_destroy/{id}": {
            "delete": {
                "tags": [
                    "The_client"
                ],
                "summary": "Supprimer un client",
                "description": "Elle se charger de supprimer un  et donc son compte",
                "operationId": "690b0a1f2b81096b219ac5f0421f71bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant du client à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " supprimé avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "L'étudiant n'existe pas !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}